+++ app/src/main/java/ru/divizdev/homefinance/data/IRepositoryWallet.kt
@@ -1,7 +1,7 @@
 package ru.divizdev.homefinance.data
 
-import ru.divizdev.homefinance.model.Wallet
+import ru.divizdev.homefinance.entities.Wallet
 
 interface IRepositoryWallet{
-    fun getWallet():Wallet
+    fun getWallet(): Wallet
 }
