+++ app/src/main/java/ru/divizdev/homefinance/entities/Wallet.kt
@@ -1,0 +1,6 @@
+package ru.divizdev.homefinance.entities
+
+//Предполагается что в кошельке все транзакции с одной главной валютой. При добавлении транзакции будет осуществляться конвертация в валюту кошелька
+class Wallet(val name: String, val mainCurrency: Currency, var listTransactions: List<Transaction>) {
+
+}
