+++ app/src/main/res/values-ru-rRU/strings.xml
@@ -8,8 +8,8 @@
     <string name="title_settings">Настройки</string>
     <string name="title_account_balance">Остаток на счетах</string>
     <string name="title_brief_overview">Краткий обзор</string>
-    <string name="title_revenue">Доход</string>
-    <string name="title_expense">Расход</string>
+    <string name="title_revenue">Доходы</string>
+    <string name="title_expense">Расходы</string>
     <string name="title_dialog_about">О приложении</string>
     <string name="title_version">Версия</string>
     <string name="pref_currency_default_value">RUB</string>
