+++ app/src/test/java/ru/divizdev/homefinance/TransactionUnitTest.kt
@@ -2,12 +2,8 @@ package ru.divizdev.homefinance
 
 import org.junit.Assert.assertEquals
 import org.junit.Test
-import ru.divizdev.homefinance.entities.Currency
-import ru.divizdev.homefinance.entities.Money
-import ru.divizdev.homefinance.entities.Transaction
-import ru.divizdev.homefinance.entities.TypeTransaction
+import ru.divizdev.homefinance.entities.*
 import ru.divizdev.homefinance.model.TransactionManager
-import ru.divizdev.homefinance.model.Wallet
 import java.math.BigDecimal
 
 /**
@@ -21,7 +17,7 @@ class TransactionUnitTest {
 
         val transactionFirst = Transaction(TypeTransaction.Expense, Money(BigDecimal.valueOf(100.405), Currency.USD))
         val transactionSecond = Transaction(TypeTransaction.Expense, Money(BigDecimal.valueOf(25.10), Currency.USD))
-        val wallet = Wallet("Sberbank", Currency.USD,  listOf(transactionFirst, transactionSecond))
+        val wallet = Wallet("Sberbank", Currency.USD, listOf(transactionFirst, transactionSecond))
         val transactionManager = TransactionManager()
 
         val money = transactionManager.calculate(wallet)
@@ -35,7 +31,7 @@ class TransactionUnitTest {
         val transactionFirst = Transaction(TypeTransaction.Revenue, Money(BigDecimal.valueOf(100.40), Currency.USD))
         val transactionSecond = Transaction(TypeTransaction.Revenue, Money(BigDecimal.valueOf(25.10), Currency.USD))
         val transactionThird = Transaction(TypeTransaction.Revenue, Money(BigDecimal.valueOf(35), Currency.USD))
-        val wallet = Wallet("Sberbank", Currency.USD,  listOf(transactionFirst, transactionSecond, transactionThird))
+        val wallet = Wallet("Sberbank", Currency.USD, listOf(transactionFirst, transactionSecond, transactionThird))
 
         val transactionManager = TransactionManager()
 
@@ -50,7 +46,7 @@ class TransactionUnitTest {
         val transactionFirst = Transaction(TypeTransaction.Revenue, Money(BigDecimal.valueOf(100.40), Currency.USD))
         val transactionSecond = Transaction(TypeTransaction.Expense, Money(BigDecimal.valueOf(25.10), Currency.USD))
         val transactionThird = Transaction(TypeTransaction.Revenue, Money(BigDecimal.valueOf(35), Currency.USD))
-        val wallet = Wallet("Sberbank", Currency.USD,  listOf(transactionFirst, transactionSecond, transactionThird))
+        val wallet = Wallet("Sberbank", Currency.USD, listOf(transactionFirst, transactionSecond, transactionThird))
         val transactionManager = TransactionManager()
 
         val money = transactionManager.calculate(wallet)
@@ -64,7 +60,7 @@ class TransactionUnitTest {
         val transactionFirst = Transaction(TypeTransaction.Revenue, Money(BigDecimal.valueOf(100.40), Currency.USD))
         val transactionSecond = Transaction(TypeTransaction.Expense, Money(BigDecimal.valueOf(25.10), Currency.USD))
         val transactionThird = Transaction(TypeTransaction.Revenue, Money(BigDecimal.valueOf(35), Currency.USD))
-        val wallet = Wallet("Sberbank", Currency.USD,  listOf(transactionFirst, transactionSecond, transactionThird))
+        val wallet = Wallet("Sberbank", Currency.USD, listOf(transactionFirst, transactionSecond, transactionThird))
         val transactionManager = TransactionManager()
 
         val moneyRevenu = transactionManager.calculate(wallet, TypeTransaction.Revenue)
